+++ cli/Valet/Brew.php
@@ -22,7 +22,7 @@ class Brew
         'php71',
         'php70',
     ];
-
+    const BREW_DISABLE_AUTO_CLEANUP = 'HOMEBREW_NO_INSTALL_CLEANUP=1';
     const LATEST_PHP_VERSION = 'php@8.1';
 
     public $cli;
@@ -163,7 +163,7 @@ class Brew
             warning('Note: older PHP versions may take 10+ minutes to compile from source. Please wait ...');
         }
 
-        $this->cli->runAsUser(trim('brew install '.$formula.' '.implode(' ', $options)), function ($exitCode, $errorOutput) use ($formula) {
+        $this->cli->runAsUser(trim(static::BREW_DISABLE_AUTO_CLEANUP.' brew install '.$formula.' '.implode(' ', $options)), function ($exitCode, $errorOutput) use ($formula) {
             output($errorOutput);
 
             throw new DomainException('Brew was unable to install ['.$formula.'].');
@@ -181,7 +181,7 @@ class Brew
         $formulas = is_array($formulas) ? $formulas : func_get_args();
 
         foreach ($formulas as $formula) {
-            $this->cli->passthru('sudo -u "'.user().'" brew tap '.$formula);
+            $this->cli->passthru(static::BREW_DISABLE_AUTO_CLEANUP.' sudo -u "'.user().'" brew tap '.$formula);
         }
     }
 
@@ -487,7 +487,7 @@ class Brew
      */
     public function uninstallFormula($formula)
     {
-        $this->cli->runAsUser('brew uninstall --force '.$formula);
+        $this->cli->runAsUser(static::BREW_DISABLE_AUTO_CLEANUP.' brew uninstall --force '.$formula);
         $this->cli->run('rm -rf '.BREW_PREFIX.'/Cellar/'.$formula);
     }
