+++ cli/Valet/Drivers/Concrete5ValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class Concrete5ValetDriver extends BasicValetDriver
 {
     /**
@@ -13,6 +15,23 @@ class Concrete5ValetDriver extends BasicValetDriver
     public function serves($sitePath, $siteName, $uri)
     {
         return file_exists($sitePath.'/concrete/config/install/base');
+    }
+
+    /**
+     * Determine if the incoming request is for a static file.
+     *
+     * @param  string  $sitePath
+     * @param  string  $siteName
+     * @param  string  $uri
+     * @return string|false
+     */
+    public function isStaticFile($sitePath, $siteName, $uri)
+    {
+        if (stripos($uri, '/application/files') === 0) {
+            return $sitePath.$uri;
+        }
+
+        return parent::isStaticFile($sitePath, $siteName, $uri);
     }
 
     /**
