+++ cli/Valet/Drivers/ContaoValetDriver.php
@@ -1,6 +1,8 @@
 <?php
 
-class StatamicV1ValetDriver extends ValetDriver
+namespace Valet\Drivers;
+
+class ContaoValetDriver extends ValetDriver
 {
     /**
      * Determine if the driver serves the request.
@@ -12,7 +14,7 @@ class StatamicV1ValetDriver extends ValetDriver
      */
     public function serves($sitePath, $siteName, $uri)
     {
-        return file_exists($sitePath.'/_app/core/statamic.php');
+        return is_dir($sitePath.'/vendor/contao') && file_exists($sitePath.'/web/app.php');
     }
 
     /**
@@ -25,14 +27,7 @@ class StatamicV1ValetDriver extends ValetDriver
      */
     public function isStaticFile($sitePath, $siteName, $uri)
     {
-        if (strpos($uri, '/_add-ons') === 0 || strpos($uri, '/_app') === 0 || strpos($uri, '/_content') === 0 ||
-            strpos($uri, '/_cache') === 0 || strpos($uri, '/_config') === 0 || strpos($uri, '/_logs') === 0 ||
-            $uri === '/admin'
-        ) {
-            return false;
-        }
-
-        if ($this->isActualFile($staticFilePath = $sitePath.$uri)) {
+        if ($this->isActualFile($staticFilePath = $sitePath.'/web'.$uri)) {
             return $staticFilePath;
         }
 
@@ -49,20 +44,17 @@ class StatamicV1ValetDriver extends ValetDriver
      */
     public function frontControllerPath($sitePath, $siteName, $uri)
     {
-        if (strpos($uri, '/admin.php') === 0) {
-            $_SERVER['SCRIPT_NAME'] = '/admin.php';
-
-            return $sitePath.'/admin.php';
+        if ($uri === '/install.php') {
+            return $sitePath.'/web/install.php';
         }
 
-        if ($uri === '/admin') {
-            $_SERVER['SCRIPT_NAME'] = '/admin/index.php';
+        if (0 === strncmp($uri, '/app_dev.php', 12)) {
+            $_SERVER['SCRIPT_NAME'] = '/app_dev.php';
+            $_SERVER['SCRIPT_FILENAME'] = $sitePath.'/app_dev.php';
 
-            return $sitePath.'/admin/index.php';
+            return $sitePath.'/web/app_dev.php';
         }
 
-        $_SERVER['SCRIPT_NAME'] = '/index.php';
-
-        return $sitePath.'/index.php';
+        return $sitePath.'/web/app.php';
     }
 }
