+++ cli/Valet/Drivers/DrupalValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class DrupalValetDriver extends ValetDriver
 {
     /**
