+++ cli/Valet/Drivers/JigsawValetDriver.php
@@ -1,6 +1,8 @@
 <?php
 
-class KatanaValetDriver extends BasicValetDriver
+namespace Valet\Drivers;
+
+class JigsawValetDriver extends BasicValetDriver
 {
     /**
      * Mutate the incoming URI.
@@ -10,7 +12,7 @@ class KatanaValetDriver extends BasicValetDriver
      */
     public function mutateUri($uri)
     {
-        return rtrim('/public'.$uri, '/');
+        return rtrim('/build_local'.$uri, '/');
     }
 
     /**
@@ -23,6 +25,6 @@ class KatanaValetDriver extends BasicValetDriver
      */
     public function serves($sitePath, $siteName, $uri)
     {
-        return file_exists($sitePath.'/katana');
+        return is_dir($sitePath.'/build_local');
     }
 }
