+++ cli/Valet/Drivers/JoomlaValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class JoomlaValetDriver extends BasicValetDriver
 {
     /**
