+++ cli/Valet/Drivers/KirbyValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class KirbyValetDriver extends ValetDriver
 {
     /**
