+++ cli/Valet/Drivers/LaravelValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class LaravelValetDriver extends ValetDriver
 {
     /**
