+++ cli/Valet/Drivers/StatamicV1ValetDriver.php
@@ -1,6 +1,8 @@
 <?php
 
-class CakeValetDriver extends ValetDriver
+namespace Valet\Drivers;
+
+class StatamicV1ValetDriver extends ValetDriver
 {
     /**
      * Determine if the driver serves the request.
@@ -12,7 +14,7 @@ class CakeValetDriver extends ValetDriver
      */
     public function serves($sitePath, $siteName, $uri)
     {
-        return file_exists($sitePath.'/bin/cake');
+        return file_exists($sitePath.'/_app/core/statamic.php');
     }
 
     /**
@@ -25,7 +27,14 @@ class CakeValetDriver extends ValetDriver
      */
     public function isStaticFile($sitePath, $siteName, $uri)
     {
-        if ($this->isActualFile($staticFilePath = $sitePath.'/webroot/'.$uri)) {
+        if (strpos($uri, '/_add-ons') === 0 || strpos($uri, '/_app') === 0 || strpos($uri, '/_content') === 0 ||
+            strpos($uri, '/_cache') === 0 || strpos($uri, '/_config') === 0 || strpos($uri, '/_logs') === 0 ||
+            $uri === '/admin'
+        ) {
+            return false;
+        }
+
+        if ($this->isActualFile($staticFilePath = $sitePath.$uri)) {
             return $staticFilePath;
         }
 
@@ -42,11 +51,20 @@ class CakeValetDriver extends ValetDriver
      */
     public function frontControllerPath($sitePath, $siteName, $uri)
     {
-        $_SERVER['DOCUMENT_ROOT'] = $sitePath.'/webroot';
-        $_SERVER['SCRIPT_FILENAME'] = $sitePath.'/webroot/index.php';
-        $_SERVER['SCRIPT_NAME'] = '/index.php';
-        $_SERVER['PHP_SELF'] = '/index.php';
+        if (strpos($uri, '/admin.php') === 0) {
+            $_SERVER['SCRIPT_NAME'] = '/admin.php';
 
-        return $sitePath.'/webroot/index.php';
+            return $sitePath.'/admin.php';
+        }
+
+        if ($uri === '/admin') {
+            $_SERVER['SCRIPT_NAME'] = '/admin/index.php';
+
+            return $sitePath.'/admin/index.php';
+        }
+
+        $_SERVER['SCRIPT_NAME'] = '/index.php';
+
+        return $sitePath.'/index.php';
     }
 }
