+++ cli/Valet/Drivers/StatamicValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class StatamicValetDriver extends ValetDriver
 {
     /**
