+++ cli/Valet/Drivers/SymfonyValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 class SymfonyValetDriver extends ValetDriver
 {
     /**
