+++ cli/Valet/Drivers/Typo3ValetDriver.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Valet\Drivers;
+
 /**
  * This driver serves TYPO3 instances (version 7.0 and up). It activates, if it
  * finds the characteristic typo3/ folder in the document root, serves both
