+++ cli/Valet/Drivers/WordPressValetDriver.php
@@ -1,6 +1,8 @@
 <?php
 
-class BedrockValetDriver extends BasicValetDriver
+namespace Valet\Drivers;
+
+class WordPressValetDriver extends BasicValetDriver
 {
     /**
      * Determine if the driver serves the request.
@@ -12,29 +14,7 @@ class BedrockValetDriver extends BasicValetDriver
      */
     public function serves($sitePath, $siteName, $uri)
     {
-        return file_exists($sitePath.'/web/app/mu-plugins/bedrock-autoloader.php') ||
-              (is_dir($sitePath.'/web/app/') &&
-               file_exists($sitePath.'/web/wp-config.php') &&
-               file_exists($sitePath.'/config/application.php'));
-    }
-
-    /**
-     * Determine if the incoming request is for a static file.
-     *
-     * @param  string  $sitePath
-     * @param  string  $siteName
-     * @param  string  $uri
-     * @return string|false
-     */
-    public function isStaticFile($sitePath, $siteName, $uri)
-    {
-        $staticFilePath = $sitePath.'/web'.$uri;
-
-        if ($this->isActualFile($staticFilePath)) {
-            return $staticFilePath;
-        }
-
-        return false;
+        return file_exists($sitePath.'/wp-config.php') || file_exists($sitePath.'/wp-config-sample.php');
     }
 
     /**
@@ -52,9 +32,7 @@ class BedrockValetDriver extends BasicValetDriver
         $_SERVER['SERVER_NAME'] = $_SERVER['HTTP_HOST'];
 
         return parent::frontControllerPath(
-            $sitePath.'/web',
-            $siteName,
-            $this->forceTrailingSlash($uri)
+            $sitePath, $siteName, $this->forceTrailingSlash($uri)
         );
     }
 
@@ -66,7 +44,7 @@ class BedrockValetDriver extends BasicValetDriver
      */
     private function forceTrailingSlash($uri)
     {
-        if (substr($uri, -1 * strlen('/wp/wp-admin')) == '/wp/wp-admin') {
+        if (substr($uri, -1 * strlen('/wp-admin')) == '/wp-admin') {
             header('Location: '.$uri.'/');
             exit;
         }
