+++ cli/Valet/Site.php
@@ -208,7 +208,10 @@ class Site
             $directory = $this->host(getcwd());
         }
 
-        $directory = str_replace('.'.$tld, '', $directory); // Remove .tld from sitename if it was provided
+        // Remove .tld from the end of sitename if it was provided
+        if (ends_with($directory, '.'.$tld)) {
+            $directory = substr($directory, 0, -(strlen('.'.$tld)));
+        }
 
         if (! $this->parked()->merge($this->links())->where('site', $directory)->count() > 0) {
             throw new DomainException("The [{$directory}] site could not be found in Valet's site list.");
@@ -409,7 +412,7 @@ class Site
             $newUrl = str_replace('.'.$oldTld, '.'.$tld, $url);
             $siteConf = $this->getSiteConfigFileContents($url, '.'.$oldTld);
 
-            if (! empty($siteConf) && strpos($siteConf, '# valet stub: proxy.valet.conf') === 0) {
+            if (! empty($siteConf) && strpos($siteConf, '# valet stub: secure.proxy.valet.conf') === 0) {
                 // proxy config
                 $this->unsecure($url);
                 $this->secure(
@@ -686,7 +689,7 @@ class Site
      */
     public function buildCertificateConf($path, $url)
     {
-        $config = str_replace('VALET_DOMAIN', $url, $this->files->get(__DIR__.'/../stubs/openssl.conf'));
+        $config = str_replace('VALET_DOMAIN', $url, $this->files->getStub('openssl.conf'));
         $this->files->putAsUser($path, $config);
     }
 
@@ -701,7 +704,7 @@ class Site
     {
         if ($siteConf === null) {
             $siteConf = $this->replaceOldLoopbackWithNew(
-                $this->files->get(__DIR__.'/../stubs/secure.valet.conf'),
+                $this->files->getStub('secure.valet.conf'),
                 'VALET_LOOPBACK',
                 $this->valetLoopback()
             );
@@ -739,7 +742,7 @@ class Site
             $siteConf = str_replace(
                 ['VALET_HOME_PATH', 'VALET_SERVER_PATH', 'VALET_STATIC_PREFIX', 'VALET_SITE', 'VALET_PHP_FPM_SOCKET', 'VALET_ISOLATED_PHP_VERSION'],
                 [VALET_HOME_PATH, VALET_SERVER_PATH, VALET_STATIC_PREFIX, $valetSite, PhpFpm::fpmSockName($phpVersion), $phpVersion],
-                $this->replaceLoopback($this->files->get(__DIR__.'/../stubs/site.valet.conf'))
+                $this->replaceLoopback($this->files->getStub('site.valet.conf'))
             );
         }
 
@@ -849,9 +852,7 @@ class Site
         }
 
         $siteConf = $this->replaceOldLoopbackWithNew(
-            $this->files->get(
-                $secure ? __DIR__.'/../stubs/secure.proxy.valet.conf' : __DIR__.'/../stubs/proxy.valet.conf'
-            ),
+            $this->files->getStub($secure ? 'secure.proxy.valet.conf' : 'proxy.valet.conf'),
             'VALET_LOOPBACK',
             $this->valetLoopback()
         );
@@ -976,7 +977,7 @@ class Site
                 str_replace(
                     'VALET_LOOPBACK',
                     $loopback,
-                    $this->files->get(__DIR__.'/../stubs/loopback.plist')
+                    $this->files->getStub('loopback.plist')
                 )
             );
