--- cli/drivers/require.php
@@ -1,30 +1,0 @@
-<?php
-
-/**
- * Basic drivers...
- */
-require_once __DIR__.'/ValetDriver.php';
-require_once __DIR__.'/BasicValetDriver.php';
-
-/**
- * Specific drivers...
- */
-require_once __DIR__.'/CraftValetDriver.php';
-require_once __DIR__.'/JigsawValetDriver.php';
-require_once __DIR__.'/KirbyValetDriver.php';
-require_once __DIR__.'/LaravelValetDriver.php';
-require_once __DIR__.'/SculpinValetDriver.php';
-require_once __DIR__.'/StatamicValetDriver.php';
-require_once __DIR__.'/StatamicV1ValetDriver.php';
-require_once __DIR__.'/SymfonyValetDriver.php';
-require_once __DIR__.'/WordPressValetDriver.php';
-require_once __DIR__.'/BedrockValetDriver.php';
-require_once __DIR__.'/ContaoValetDriver.php';
-require_once __DIR__.'/KatanaValetDriver.php';
-require_once __DIR__.'/CakeValetDriver.php';
-require_once __DIR__.'/JoomlaValetDriver.php';
-require_once __DIR__.'/DrupalValetDriver.php';
-require_once __DIR__.'/Concrete5ValetDriver.php';
-require_once __DIR__.'/Typo3ValetDriver.php';
-require_once __DIR__.'/NeosValetDriver.php';
-require_once __DIR__.'/Magento2ValetDriver.php';
