+++ cli/includes/ValetDriver.php
@@ -1,0 +1,7 @@
+<?php
+
+use Valet\Drivers\ValetDriver as RealValetDriver;
+
+abstract class ValetDriver extends RealValetDriver
+{
+}
