+++ cli/includes/require-drivers.php
@@ -1,0 +1,10 @@
+<?php
+
+require_once './cli/Valet/Drivers/ValetDriver.php';
+require_once './cli/includes/ValetDriver.php';
+
+foreach (scandir('./cli/Valet/Drivers') as $file) {
+    if (substr($file, 0, 1) !== '.') {
+        require_once './cli/Valet/Drivers/'.$file;
+    }
+}
