+++ composer.json
@@ -11,13 +11,21 @@
         {
             "name": "Adam Wathan",
             "email": "adam.wathan@gmail.com"
+        },
+        {
+            "name": "Matt Stauffer",
+            "email": "matt@tighten.co"
         }
     ],
     "autoload": {
         "files": [
             "cli/includes/compatibility.php",
             "cli/includes/facades.php",
-            "cli/includes/helpers.php"
+            "cli/includes/helpers.php",
+            "cli/includes/ValetDriver.php",
+            "tests/UsesNullWriter.php",
+            "tests/BaseApplicationTestCase.php",
+            "tests/Drivers/BaseDriverTestCase.php"
         ],
         "psr-4": {
             "Valet\\": "cli/Valet/"
@@ -27,9 +35,11 @@
         "php": "^7.0|^8.0",
         "illuminate/container": "~5.1|^6.0|^7.0|^8.0|^9.0",
         "mnapoli/silly": "^1.0",
+        "symfony/console": "^3.0|^4.0|^5.0|^6.0",
         "symfony/process": "^3.0|^4.0|^5.0|^6.0",
         "tightenco/collect": "^5.3|^6.0|^7.0|^8.0",
-        "guzzlehttp/guzzle": "^6.0|^7.4"
+        "guzzlehttp/guzzle": "^6.0|^7.4",
+        "symfony/event-dispatcher": "^3.0|^4.0|^5.0|^6.0"
     },
     "require-dev": {
         "mockery/mockery": "^1.2.3",
